+++ lib/reach/check/changed.ex
@@ -4,7 +4,7 @@ defmodule Reach.Check.Changed do
   alias Reach.Check.Architecture
   alias Reach.Check.Changed.Function, as: ChangedFunction
   alias Reach.Check.Changed.Result
-  alias Reach.CloneAnalysis
+  alias Reach.Evidence.CloneAnalysis
   alias Reach.Config
   alias Reach.IR
   alias Reach.IR.Helpers, as: IRHelpers